Moves some analysis files only used for tests to test-data and makes sure they are not deleted

+++ test/main.cc
@@ -1,3 +1,4 @@
+#include <cctype>
 #define CATCH_CONFIG_RUNNER
 #include <catch2/catch.hpp>
 #include "spdlog/common.h"
@@ -24,8 +25,12 @@ int main( int argc, char* argv[] ) {
   int result = Catch::Session().run( argc, argv );
 
   // Remove all analyzed-data files.
-  for (const auto& entry : std::filesystem::directory_iterator("test_data/data/analysis/")) 
-    std::filesystem::remove_all(entry.path());
+  std::cout << "CLEAN UP" << std::endl;
+  for (const auto& entry : std::filesystem::directory_iterator("test_data/data/analysis/")) {
+    std::cout << entry.path().filename() << std::endl;
+    if (std::isdigit(entry.path().filename().string()[0]))
+      std::filesystem::remove_all(entry.path());
+  }
 
   return result;
 }
